Changed pathfinder to use less memory. Added IP Ban command. Added Machine Ban command. Done a lot more shit I cannot remember.

+++ src/main/java/com/eu/habbo/messages/incoming/floorplaneditor/FloorPlanEditorSaveEvent.java
@@ -68,8 +68,8 @@ public class FloorPlanEditorSaveEvent extends MessageHandler
 
             if(layout != null)
             {
-                layout.setDoorX(doorX);
-                layout.setDoorY(doorY);
+                layout.setDoorX((short) doorX);
+                layout.setDoorY((short) doorY);
                 layout.setDoorDirection(doorRotation);
                 layout.setHeightmap(map);
                 layout.parse();
